+++ Silhouette/src/StyleSheets/BaseComponents/AtRule.java
@@ -1,6 +1,6 @@
-package CSS.BaseComponents;
+package StyleSheets.BaseComponents;
 
-import CSS.BaseComponents.Implementation.Statement;
+import StyleSheets.BaseComponents.Implementation.Statement;
 
 import java.util.ArrayList;
 import java.util.Arrays;
