+++ Silhouette/src/StyleSheets/BaseComponents/Implementation/Statement.java
@@ -1,4 +1,4 @@
-package CSS.BaseComponents.Implementation;
+package StyleSheets.BaseComponents.Implementation;
 
 import java.util.ArrayList;
