+++ Silhouette/src/StyleSheets/GridChild.java
@@ -1,4 +1,4 @@
-package CSS;
+package StyleSheets;
 
 public class GridChild {
     // Placing Grid Items
